refactor: remove pkg folder
