+++ .gitignore
@@ -20,5 +20,4 @@ test
 *.csv
 
 /volume*
-/wallet*
